pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "golden-hour-dispatch-env"
version = "0.1.0"
description = "An OpenEnv emergency dispatch environment for Indian 108/102 ambulance allocation."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Codex + User" }
]
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "openai>=2.7.2,<3.0.0",
  "openenv-core[core]>=0.2.2,<0.4.0",
  "pydantic>=2.8.0,<3.0.0",
  "uvicorn>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["golden_hour_dispatch_env*", "server*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [
  ".git",
  ".submit-venv",
  ".uv-cache",
  ".uv-python",
  ".venv",
  "golden-hour-dispatch-submission",
  "submission",
]
addopts = "-p no:cacheprovider"
